+++ spec/eve_online/esi/character_online_spec.rb
@@ -1,0 +1,29 @@
+require 'spec_helper'
+
+describe EveOnline::ESI::CharacterOnline do
+  let(:options) { { token: 'token123', character_id: 12_345_678 } }
+
+  subject { described_class.new(options) }
+
+  specify { expect(subject).to be_a(EveOnline::ESI::Base) }
+
+  specify { expect(described_class::API_ENDPOINT).to eq('https://esi.tech.ccp.is/v2/characters/%s/online/?datasource=tranquility') }
+
+  describe '#initialize' do
+    its(:token) { should eq('token123') }
+
+    its(:parser) { should eq(JSON) }
+
+    its(:character_id) { should eq(12_345_678) }
+  end
+
+  describe '#scope' do
+    specify { expect(subject.scope).to eq('esi-location.read_online.v1') }
+  end
+
+  describe '#url' do
+    specify do
+      expect(subject.url).to eq('https://esi.tech.ccp.is/v2/characters/12345678/online/?datasource=tranquility')
+    end
+  end
+end